requirements.txt
crewai==0.165.1
requests==2.31.0
python-dotenv==1.0.0 
fastapi==0.116.1
uvicorn==0.35.0
pybase64==1.4.2
beautifulsoup4==4.13.5
google-api-python-client==2.178.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
pydantic==2.11.7
pysqlite3-binary==0.5.4